requirements.txt
tensorflow==2.16.1
streamlit==1.25.0
numpy
pandas
Pillow